Rotate command Now merged into one command /rotate [Left|Right]
